verification/requirements.txt
cocotbext-uart
pytest
bitstring~=3.1
PyYAML
Deprecated
